Move globals.js up for simpler ref by non-App modules
